LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/animation-range-normal-matches-cover-expected.txt


FAIL Changing the animation range updates the play state assert_approx_equals: values do not match for "currentTime at start of normal range" expected 0 +/- 0.125 but got 100
